added constraints, added german translations of errors and implemented half coin values

+++ app/models/coin/WorkEntry.scala
@@ -7,7 +7,7 @@ import scalikejdbc._
 
 import scala.concurrent.duration._
 
-case class WorkEntry(id: WorkEntryId, personId: PersonId, area: String, task: String, description: String, timeSpent: Long, coins: Int, dateDone: DateTime) {
+case class WorkEntry(id: WorkEntryId, personId: PersonId, area: String, task: String, description: String, timeSpent: Long, coins: Double, dateDone: DateTime) {
   val prettyTimeSpent: String = s"0${timeSpent.minutes.toHours}:".takeRight(3) + s"0${timeSpent.minutes.minus(timeSpent.minutes.toHours.hours).toMinutes}".takeRight(2)
 }
 
@@ -19,7 +19,7 @@ object WorkEntry extends SQLSyntaxSupport[WorkEntry]{
   val w: scalikejdbc.QuerySQLSyntaxProvider[scalikejdbc.SQLSyntaxSupport[WorkEntry], WorkEntry] = WorkEntry.syntax
   val wc: scalikejdbc.ColumnName[WorkEntry] = WorkEntry.column
 
-  def create(personId: PersonId, area: String, task: String, description: String, timeSpent: Long, coins: Int, dateDone:DateTime)(implicit s: DBSession = AutoSession): Long = {
+  def create(personId: PersonId, area: String, task: String, description: String, timeSpent: Long, coins: Double, dateDone:DateTime)(implicit s: DBSession = AutoSession): Long = {
     withSQL {
       insert.into(WorkEntry).namedValues(
         wc.personId -> personId,
@@ -50,6 +50,6 @@ object WorkEntry extends SQLSyntaxSupport[WorkEntry]{
   }
 
   def apply(r: ResultName[WorkEntry])(rs: WrappedResultSet): WorkEntry = {
-    new WorkEntry(rs.long(r.id), rs.long(r.personId), rs.string(r.area), rs.string(r.task), rs.string(r.description), rs.long(r.timeSpent), rs.int(r.coins), rs.jodaDateTime(r.dateDone))
+    new WorkEntry(rs.long(r.id), rs.long(r.personId), rs.string(r.area), rs.string(r.task), rs.string(r.description), rs.long(r.timeSpent), rs.double(r.coins), rs.jodaDateTime(r.dateDone))
   }
 }
