+++ conf/evolutions/default/2.sql
@@ -12,7 +12,7 @@ CREATE TABLE work_entries (
     task varchar(255) NOT NULL,
     description varchar(255) NOT NULL,
     time_spent int NOT NULL,
-    coins int NOT NULL,
+    coins double NOT NULL,
     date_done date NOT NULL
 );
